Split out ConceptualBehaviourTree functionality into helper methods
This is the first of a set of changes to make the workflow with the models simpler. The models should be just representing the data of the behaviour tree whilst all edits should be made either via the extension methods defined in the BehaviourTrees.Model assembly or any methods that editors might define themselves.

The methods to validate and update the values of the node model will be implemented and expanded on later.